requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
GeoAlchemy2==0.14.2
geopy==2.4.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Pillow==10.1.0
python-dotenv==1.0.0